+++ R/SpotVA1.R
@@ -23,7 +23,7 @@
 #'
 #' @seealso \code{SpotVA2} in \pkg{FSAdata}.
 #'
-#' @source Extracted from Table 1 in Chapter 8 of the \href{http://ww2.odu.edu/sci/cqfe/}{Virginia Marine Resources Commissions's} Final Report on finfish ageing, 2002.
+#' @source Extracted from Table 1 in Chapter 8 (Spot) of the VMRC Final Report on Finfish Ageing, 2002 by the \href{http://ww2.odu.edu/sci/cqfe/}{Center for Quantitative Fisheries Ecology} at Old Dominion University.
 #'
 #' @keywords datasets
 #'